ECMS-3356 Allow the Multithreaded service of Open Office

+++ core/viewer/src/main/java/org/exoplatform/ecm/REST/viewer/PDFViewerRESTService.java
@@ -28,7 +28,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.Serializable;
-import java.net.ConnectException;
 
 import javax.imageio.ImageIO;
 import javax.jcr.Node;
@@ -38,6 +37,7 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
 
+import org.artofsolving.jodconverter.office.OfficeException;
 import org.exoplatform.container.ExoContainer;
 import org.exoplatform.container.ExoContainerContext;
 import org.exoplatform.services.cache.CacheService;
@@ -56,8 +56,6 @@ import org.icepdf.core.exceptions.PDFSecurityException;
 import org.icepdf.core.pobjects.Document;
 import org.icepdf.core.pobjects.Page;
 import org.icepdf.core.util.GraphicsRenderingHints;
-
-import com.artofsolving.jodconverter.openoffice.connection.OpenOfficeException;
 
 /**
  * Created by The eXo Platform SARL
@@ -279,7 +277,7 @@ public class PDFViewerRESTService implements ResourceContainer {
       Node contentNode = currentNode.getNode("jcr:content");
       String mimeType = contentNode.getProperty("jcr:mimeType").getString();
       InputStream input = new BufferedInputStream(contentNode.getProperty("jcr:data").getStream());
-      // Create temp file to store data of nt:file node
+      // Create temp file to store converted data of nt:file node
       if (name.indexOf(".") > 0) name = name.substring(0, name.lastIndexOf("."));
       content = File.createTempFile(name + "_tmp", ".pdf");
       /*
@@ -293,19 +291,22 @@ public class PDFViewerRESTService implements ResourceContainer {
         read(input, new BufferedOutputStream(new FileOutputStream(content)));
       } else {
         OutputStream out = new BufferedOutputStream((new FileOutputStream(content)));
+        // create temp file to store original data of nt:file node
+        File in = File.createTempFile(name + "_tmp", null);
+        read(input, new BufferedOutputStream(new FileOutputStream(in)));
         try {
-          jodConverter_.convert(input, extension, out, "pdf");
-        } catch(ConnectException connection) {
-          content.delete();
-          if (LOG.isErrorEnabled()) {
-            LOG.error("Cannot open connection to OpenOffice Service");
+          boolean success = jodConverter_.convert(in, content, "pdf");
+          // If the converting was failure then delete the content temporary file
+          if (!success) {
+            content.delete();
           }
-        } catch(OpenOfficeException connection) {
+        } catch (OfficeException connection) {
           content.delete();
           if (LOG.isErrorEnabled()) {
-            LOG.error("Exception when using OpenOffice Service");
+            LOG.error("Exception when using Office Service");
           }
         } finally {
+          in.delete();
           out.flush();
           out.close();
         }
